refactor(control): the harness owns reading the engine, the recorder owns the file
CONTROL is the one module that reads the game's globals, for --dump-state, and
so the one that includes C_EXTERN.H and with it the DEFINES.H aggregation. A
second module doing the same doubles a dependency the build graph deliberately
holds down: check-build-graph rule 2 counted 58 translation units against a
ceiling of 57, and says plainly that raising it is not a fix.

So the state digest, the values behind it, the keyframe, the state description
and the savegame handling move to CONTROL, and RECORD keeps the stream, the
format, the bindings and the clock. It reads no game global at all now, which is
the same line RECORD_FORMAT is already cut along.

Three things fall out of putting them together:

- The save-context dance is in one place. SaveGame writes "the current game", so
  writing a snapshot anywhere else means briefly becoming that game, borrowing
  the player's name, path and version and putting them back. That was written
  twice, once for the snapshot and once around the mid-session load.
- The digest, the named values and the keyframe are three views of one walk over
  the state, and now share it.
- The binding digest keeps its own FNV. It hashes the tables the file carries,
  not the game, and the two should be able to change without disturbing each
  other.

Behaviour-neutral, and checked as such rather than assumed: three recordings
replay to the same tick with the same hashes as before, including the one that
diverges at 1492 with recorded d85f57f81fb64ee8 against replayed
119fce4459e12a57. A digest that shifted by a byte would have invalidated every
recording made so far.

The build graph reports 57 unchanged under the CI configuration.

Also adds the three missing CLI contract cases. Every flag in the table needs
one, and --ignore-focus, --record and --replay had none, so the suite failed on
coverage rather than on behaviour.
